+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.1.1 [unreleased]
 
 - Filter passwords in urls
+- Added `Logstop.scrub` method
 
 ## 0.1.0